avoiding implicit type conversions

+++ src/mainwindow.cpp
@@ -108,7 +108,7 @@ void MainWindow::on_Particle_number_slider_valueChanged(int particle_number_) {
 void MainWindow::on_Particle_diameter_slider_valueChanged(
 		int particle_diameter_mm) {
 	//ui->Particle_diameter_value->setNum(particle_diameter_mm);
-	const float r = particle_diameter_mm / 1000. / 2.; // int [mm] --> float [m], diameter --> radius
+	const float r = static_cast<float>(particle_diameter_mm) / 1000.0f / 2.0f; // int [mm] --> float [m], diameter --> radius
 	Particle::setUniformRadius(r);
 	updateGUIcontrols();
 
@@ -142,7 +142,7 @@ void MainWindow::on_cells_Nz_SpinBox_valueChanged(int Nz_) {
 }
 
 void MainWindow::on_Drag_coefficient_slider_valueChanged(int drag100) {
-	const float Cd = drag100 / 100.; // value of integer slider is converted to float
+	const float Cd = drag100 / 100.0f; // value of integer slider is converted to float
 	Particle::setCd(Cd); // setting static data member
 	updateGUIcontrols();
 }
@@ -204,12 +204,12 @@ void MainWindow::updateGUIcontrols() {
 	ui->Particle_number_slider->setValue(scene.getNumberOfParticles());
 	ui->Particle_number_value->setText(QString::number(scene.getNumberOfParticles()));
 
-	int radius_mm = std::round(Particle::getUniformRadius() * 2. * 1000.);
+	int radius_mm = std::round(Particle::getUniformRadius() * 2.0f * 1000.0f);
 	ui->Particle_diameter_slider->setValue(radius_mm);
 	ui->Particle_diameter_value->setText(QString::number(radius_mm) + " mm");
 
 	ui->Drag_coefficient_value->setText(QString::number(Particle::getCd()));
-	ui->Drag_coefficient_slider->setValue(int(Particle::getCd() * 100.)); // float internal value is transformed to int on the slider
+	ui->Drag_coefficient_slider->setValue(static_cast<int>(Particle::getCd() * 100.0f)); // float internal value is transformed to int on the slider
 
 	if (scene.getGeometry() == Geometry::hourglass_with_removable_orifice) {
 		ui->openOrificeButton->show();
